Added basic functionality to choose to upload files to IIIF server
