+++ specifyweb/frontend/js_src/lib/attachmentplugin.js
@@ -6,7 +6,8 @@ var _ = require('underscore');
 
 var api         = require('./specifyapi.js');
 var UIPlugin    = require('./uiplugin.js');
-var attachments = require('./attachments.js');
+var attachmentserverprivate = require('./attachmentserverbase.js');
+var attachmentserverpublic = require('./attachmentserverpublic.js');
 
 module.exports =  UIPlugin.extend({
         __name__: "AttachmentsPlugin",
@@ -16,7 +17,7 @@ module.exports =  UIPlugin.extend({
         },
         render: function() {
             var self = this;
-            if (!attachments) {
+            if (!attachmentserverprivate) {
                 self.$el.replaceWith('<div>Attachment server unavailable.</div>');
                 return this;
             }
@@ -34,15 +35,16 @@ module.exports =  UIPlugin.extend({
             return this;
         },
         addAttachment: function() {
-            this.$el.append('<form enctype="multipart/form-data"><input type="file" name="file"></form>');
-            this.$('input').click();
+            this.$el.append('<form enctype="multipart/form-data"><input type="checkbox" id="isPublicImage">  Is this a public image file?<br><input type="file" name="file"></form>');
         },
         fileSelected: function(evt) {
             var files = this.$(':file').get(0).files;
             if (files.length === 0) return;
-            this.startUpload(files[0]);
+
+            var isPublicImage = this.$('#isPublicImage').get(0).checked;
+            this.startUpload(files[0], isPublicImage);
         },
-        startUpload: function(file) {
+        startUpload: function(file, isPublicImage) {
             var self = this;
 
             self.progressBar = $('<div class="attachment-upload-progress">').progressbar();
@@ -52,7 +54,10 @@ module.exports =  UIPlugin.extend({
                 .append(self.progressBar)
                 .dialog({ modal:true });
 
-            attachments.uploadFile(file, function(progressEvt) {
+            if (isPublicImage) { var attachmentserver = attachmentserverpublic;}
+            else {var attachmentserver = attachmentserverprivate;}
+
+            attachmentserver.uploadFile(file, function(progressEvt) {
                 self.uploadProgress(progressEvt);
             }).done(function(attachment) {
                 self.uploadComplete(attachment);
@@ -80,15 +85,19 @@ module.exports =  UIPlugin.extend({
             var self = this;
             self.$el.empty().append('<div class="specify-attachment-display">');
 
-            attachments.getThumbnail(attachment).done(function(img) {
+            if (attachment.attributes.ispublic) { var attachmentserver = attachmentserverpublic;}
+            else {var attachmentserver = attachmentserverprivate;}
+
+            attachmentserver.getThumbnail(attachment).done(function(img) {
                 $('<a>').append(img).appendTo(self.$('.specify-attachment-display'));
             });
         },
         openOriginal: function(evt) {
             evt.preventDefault();
             this.model.rget('attachment', true).done(function(attachment) {
-                attachments.openOriginal(attachment);
+                if (attachment.get('ispublic')) { var attachmentserver = attachmentserverpublic;}
+                else {var attachmentserver = attachmentserverprivate;}
+                attachmentserver.openOriginal(attachment);
             });
         }
     }, { pluginsProvided: ['AttachmentPlugin'] });
-
